backend/requirements.txt
# requirements.txt - Face Recognition Project

# Core dependencies
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.2

# Computer vision
opencv-python==4.8.1.78
pillow==10.1.0
imageio==2.31.1

# Face recognition
dlib==19.24.1
face-recognition==1.3.0

# Deep learning
tensorflow==2.13.1
mtcnn==0.1.1

# Web framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Utilities
python-dotenv==1.0.0
click==8.1.7

# Optional: Image processing and visualization
matplotlib==3.7.2
